refactor(frontend): split code
use tanstack query
